Added flags for CMake Debug and GCC 4.8 support

+++ main/license.cc
@@ -954,7 +954,7 @@ int SendData(int sockfd, const char* hwid)
     // send the headers
     snprintf(buffer, STRLONG, "POST %s HTTP/1.0\n", LICENSE_PATH);
     bytes = write(sockfd, buffer, strlen(buffer));
-    snprintf(buffer, STRLONG, "Content-Length: %ld\n\n", strlen(varsbuff));
+    snprintf(buffer, STRLONG, "Content-Length: %ld\n\n", (long) strlen(varsbuff));
     bytes = write(sockfd, buffer, strlen(buffer));
 
     // send the body
